refactor(bdpan): 将 Bdpan 弹窗组件迁移至 base/ 并去除目录层级
独立文件夹（BdpanImportFilePicker/index.tsx 等）只有单文件，
增加了目录层级却没有带来模块化收益，改为直接在 base/ 下放具名文件，
与 base/ 目录下其他单文件组件的组织风格保持一致。
